refactor: compiler: change directory structure
