+++ src/frontend/compiler/src/Parser/Ast.hs
@@ -7,7 +7,7 @@ module Ast where
 data BinOp = Multiple | Divide   -- (*) (/)
            | Plus     | Minus    -- (+) (-)
            | Ls | LE  | Gr | GE  -- (<) (<=) (>) (>=)
-           | Eq | NEq | In       -- (= ==) (!= <>) (IN)
+           | Eq | NE  | In       -- (= ==) (!= <>) (IN)
            | And| Or             -- (AND) (OR)
            deriving (Show, Eq)
